# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "api/user/add-comment": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Add a comment under a post",
                "description": "Add a comment under a post",
                "operationId": "ad6daf1fa21e07d6159f2559768bb050",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "user_id",
                                    "comment"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment added Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/user/comments/{post_id}": {
            "get": {
                "tags": [
                    "Commment"
                ],
                "summary": "Get comments from post",
                "description": "Get comments from post",
                "operationId": "7b6fea5f10ca0dee45112a4a532d920c",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to get comments"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/credential/education": {
            "post": {
                "tags": [
                    "Credential"
                ],
                "summary": "user update education credentials",
                "description": "user update education credentials",
                "operationId": "702a0a41cecd0eba028037bfdd745d66",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "school",
                                    "primary_major",
                                    "secondary_major",
                                    "degree_type",
                                    "graduation_type"
                                ],
                                "properties": {
                                    "school": {
                                        "type": "nullable|string"
                                    },
                                    "primary_major": {
                                        "type": "nullable|string"
                                    },
                                    "secondary_major": {
                                        "type": "nullable|string"
                                    },
                                    "degree_type": {
                                        "type": "nullable|string"
                                    },
                                    "graduation_year": {
                                        "type": "nullable|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/credential/education/{id}": {
            "post": {
                "tags": [
                    "Credential"
                ],
                "summary": "user update education credentials",
                "description": "user update education credentials",
                "operationId": "fe06c10b9256d3d86cfc446822e31dbd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "school",
                                    "primary_major",
                                    "secondary_major",
                                    "degree_type",
                                    "graduation_type"
                                ],
                                "properties": {
                                    "school": {
                                        "type": "nullable|string"
                                    },
                                    "primary_major": {
                                        "type": "nullable|string"
                                    },
                                    "secondary_major": {
                                        "type": "nullable|string"
                                    },
                                    "degree_type": {
                                        "type": "nullable|string"
                                    },
                                    "graduation_year": {
                                        "type": "nullable|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/credential/education/{user_id}": {
            "get": {
                "tags": [
                    "Credential"
                ],
                "summary": "Get user education details",
                "description": "Get user education details",
                "operationId": "e12e3d7e5c4e73d17ebe6686435aa6c9",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/deleteEducation/{id}": {
            "delete": {
                "tags": [
                    "Delete"
                ],
                "summary": "Delete data by id",
                "description": "Delete data by id",
                "operationId": "827b2bc0ad5fc8ef2485b35d45199740",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "type": "integer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}